+++ CHANGES.md
@@ -24,6 +24,7 @@ Changes in 3.3
 - Check for PHP gd extension in diagnostic.php & INSTALL.md
 - Add $FileUploadsPath & $PNGQuantPath optional configuration variables in INSTALL.md & Warehouse.php
 - Add Image resize and compress class in classes/ImageResizeGD.php
+- Add ImageUpload() function in FileUpload.fnc.php
 
 Changes in 3.2
 --------------